update to separate cli command not mcp tool
