+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Commands/SharedOptions.cs
@@ -9,7 +9,6 @@ using Azure.Sdk.Tools.Cli.Tools.Example;
 using Azure.Sdk.Tools.Cli.Tools.TypeSpec;
 using Azure.Sdk.Tools.Cli.Tools.Verify;
 using Azure.Sdk.Tools.Cli.Tools.APIView;
-using Azure.Sdk.Tools.Cli.Tools.CodeCustomization;
 using Azure.Sdk.Tools.Cli.Tools.Package.Samples;
 using Azure.Sdk.Tools.Cli.Tools.Core;
 using Azure.Sdk.Tools.Cli.Tools.Config;
@@ -49,7 +48,7 @@ namespace Azure.Sdk.Tools.Cli.Commands
             typeof(TypeSpecPublicRepoValidationTool),
             typeof(TypeSpecAuthoringTool),
             typeof(APIViewReviewTool),
-            typeof(AddressFeedbackTool),
+            typeof(DelegateApiViewFeedbackTool),
             typeof(VerifySetupTool),
             typeof(TestTool),
             typeof(ListCommandTool),
